fit deposit-cap fields in padding and enforce cap

+++ programs/velocity/src/lib.rs
@@ -1222,7 +1222,7 @@ pub mod velocity {
 
     pub fn update_spot_market_withdraw_circuit_breaker(
         ctx: Context<AdminUpdateSpotMarket>,
-        withdraw_circuit_breaker_pct: u32,
+        withdraw_circuit_breaker_pct: u16,
     ) -> Result<()> {
         handle_update_spot_market_withdraw_circuit_breaker(ctx, withdraw_circuit_breaker_pct)
     }
@@ -1230,7 +1230,7 @@ pub mod velocity {
     pub fn update_spot_market_deposit_cap(
         ctx: Context<AdminUpdateSpotMarket>,
         deposit_guard_threshold: u64,
-        max_deposit_pct_per_day: u32,
+        max_deposit_pct_per_day: u16,
     ) -> Result<()> {
         handle_update_spot_market_deposit_cap(ctx, deposit_guard_threshold, max_deposit_pct_per_day)
     }
